<REFACTOR> move media start/end from Application to Media
